+++ screens/Ratings/DailyCardRatingsBar.tsx
@@ -4,7 +4,7 @@ import { Button } from "react-native-paper";
 import { blue100 } from "react-native-paper/lib/typescript/styles/colors";
 import { auth } from "../../firebase/firebaseConfig";
 
-export const RatingCardRatingsBar = ({
+export const DailyCardRatingsBar = ({
   ratingParameter,
   selection,
   setSelection,
