+++ screens/Ratings/RatingsStack.tsx
@@ -1,18 +1,15 @@
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { RatingsProvider } from "../../contexts/RatingsContext";
-import {RatingsScreen} from "./RatingsScreen";
+import { DailysScreen } from "./DailysScreen";
 
 const Stack = createNativeStackNavigator();
 
 export const RatingsStack = () => {
   return (
     <RatingsProvider>
-        <Stack.Navigator>
-          <Stack.Screen
-            name={"RatingsScreen"}
-            component={RatingsScreen}
-          />
-        </Stack.Navigator>
+      <Stack.Navigator>
+        <Stack.Screen name={"RatingsScreen"} component={DailysScreen} />
+      </Stack.Navigator>
     </RatingsProvider>
   );
 };
